+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:65180/25748590-5902-4eb6-a2a4-cd577b49cbf5?"}
+{"mongoUri":"mongodb://127.0.0.1:65314/36688c11-95b8-4629-9778-b38d8b0d5424?"}
