+++ src/data/usecases/index.ts
@@ -1,2 +1,2 @@
-export * from './authentication/db-authentication'
-export * from './add-account/db-add-account'
+export * from './account/authentication/db-authentication'
+export * from './account/add-account/db-add-account'
